# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Transit Tracking API",
    "version": "1.0.0",
    "description": "API for managing and tracking trips in the Rwanda transit system."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/trips": {
      "get": {
        "summary": "Get all trips",
        "tags": ["Trips"],
        "responses": {
          "200": {
            "description": "List of trips",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new trip",
        "tags": ["Trips"],
        "requestBody": {
          "description": "Trip object to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/trips/{id}": {
      "get": {
        "summary": "Get trip by ID",
        "tags": ["Trips"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the trip to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "404": {
            "description": "Trip not found"
          }
        }
      }
    },
    "/trips/{id}/board": {
      "post": {
        "summary": "Board a trip (decrement available seats)",
        "tags": ["Trips"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the trip to board",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "User boarding the trip (optional userId)",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer",
                    "format": "int32",
                    "description": "ID of the user boarding"
                  }
                }
              
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Boarded successfully, seat count decremented",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "availableSeats": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No seats available or invalid request"
          },
          "404": {
            "description": "Trip not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Trip": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "bus_id": { "type": "integer", "format": "int32" },
          "route_id": { "type": "integer", "format": "int32" },
          "departure_time": { "type": "string", "format": "date-time" },
          "available_seats": { "type": "integer" }
        },
        "required": ["id", "bus_id", "route_id", "departure_time", "available_seats"]
      },
      "TripInput": {
        "type": "object",
        "properties": {
          "bus_id": { "type": "integer", "format": "int32" },
          "route_id": { "type": "integer", "format": "int32" },
          "departure_time": { "type": "string", "format": "date-time" },
          "available_seats": { "type": "integer" }
        },
        "required": ["bus_id", "route_id", "departure_time", "available_seats"]
      }
    }
  },
  "tags": [
    {
      "name": "Trips",
      "description": "Operations related to trip management"
    }
  ]
}